fix : 수익률 계산에서 오버플로우 문제 수정
- 오버플로우를 피하기 위해서 profit의 자료형을 수정

+++ src/main/java/lotto/model/judge/LottoJudge.java
@@ -53,7 +53,7 @@ public class LottoJudge {
 
     public LottosResult getResult(){
         LottosResult lottosResult = new LottosResult();
-        int profit = 0;
+        long profit = 0;
         for(LottoWinningInfo rank : LottoWinningInfo.values()){
             lottosResult.addRow(rank.getCondition(), rank.getWinnings(), rankCount[rank.getRank().getValue()]);
             profit += rankCount[rank.getRank().getValue()] * rank.getWinnings();
